feat: auto-infer typed enum params from int32 in idiomgen
C headers declare enum params as bare `int`, so the spec records
`int32`. This made examples require ugly casts like
`mgr.DefaultSensor(int32(sensor.Accelerometer))`.

Fix: after collecting value enums, build a set of their Go names.
When a function param has type int32 and PascalCase(paramName) matches
a value enum GoName, use the typed enum instead. Also supports
explicit overlay override via `param_types` field on FuncOverlay.

Affects 9 params across 7 modules (sensor, asset, config, bitmap,
logging, nativewindow, neuralnetworks). Also fixes capiArg to not
prefix scalar Go types with `capi.` (e.g. int32 → int32(), not
capi.int32()).

Examples updated to remove int32() casts.

+++ examples/asset/overview/main.go
@@ -108,7 +108,7 @@ func main() {
 
 	if mgr != nil {
 		// Open a text file from the APK's assets/ directory.
-		a := mgr.Open("data/config.json", int32(asset.Streaming))
+		a := mgr.Open("data/config.json", asset.Streaming)
 		defer a.Close()
 
 		// Query total size. Length returns off_t (32-bit on older ABIs),
@@ -141,7 +141,7 @@ func main() {
 		// With Buffer mode, the entire asset is accessible through a
 		// direct pointer without calling Read:
 		//
-		//   a := mgr.Open("data/config.json", int32(asset.Buffer))
+		//   a := mgr.Open("data/config.json", asset.Buffer)
 		//   ptr := a.Buffer()  // unsafe.Pointer to the raw data
 		//   data := unsafe.Slice((*byte)(ptr), a.Length64())
 	}
